Add fade out animation for mobile menu using hook.
