--- src/components/FadeOutTest/index.js
@@ -1,2 +1,0 @@
-export * from './FadeOutTest';
-export { default } from './FadeOutTest';
