+++ src/components/MobileMenu/MobileMenu.js
@@ -6,59 +6,72 @@ import UnstyledButton from '../UnstyledButton';
 import VisuallyHidden from '../VisuallyHidden';
 import { Dialog } from '@headlessui/react';
 import { X } from 'react-feather';
-import { motion, AnimatePresence } from 'framer-motion';
 
 function MobileMenu({ isOpen, toggle }) {
+	const [shouldRender, setShouldRender] = React.useState(isOpen);
+
+	React.useEffect(() => {
+		if (isOpen) {
+			setShouldRender(true);
+		}
+	}, [isOpen]);
+
+	function handleAnimationEnd() {
+		setShouldRender(isOpen);
+	}
+
 	return (
-		<AnimatePresence>
-			{isOpen && (
-				<Menu open={isOpen} onClose={toggle}>
-					<Backdrop
-						initial={{ opacity: 0 }}
-						animate={{ opacity: 0.4 }}
-						exit={{ opacity: 0 }}
-						transition={{ type: 'tween', duration: '0.3' }}
-					/>
-					<PanelWrapper
-						initial={{ y: '-100%' }}
-						animate={{ y: '0%' }}
-						exit={{ y: '-100%' }}
-						transition={{ type: 'tween', duration: '0.3' }}
-					>
-						<Panel>
-							<Logo>
-								<Link href="/">JimmyJim</Link>
-							</Logo>
-							<Spacer size={38} />
-							<NavigationList>
-								<ListItem>
-									<Link href="/about">關於我</Link>
-								</ListItem>
-								<ListItem>
-									<Link href="/blog">文章</Link>
-								</ListItem>
-							</NavigationList>
-							<IconWrapper onClick={toggle}>
-								<X />
-								<VisuallyHidden>Close mobile menu</VisuallyHidden>
-							</IconWrapper>
-						</Panel>
-					</PanelWrapper>
-				</Menu>
-			)}
-		</AnimatePresence>
+		<Menu open={shouldRender} onClose={toggle}>
+			<Backdrop />
+			<Panel isOpen={isOpen} onAnimationEnd={handleAnimationEnd}>
+				<Logo>
+					<Link href="/">JimmyJim</Link>
+				</Logo>
+				<Spacer size={38} />
+				<NavigationList>
+					<ListItem>
+						<Link href="/about">關於我</Link>
+					</ListItem>
+					<ListItem>
+						<Link href="/blog">文章</Link>
+					</ListItem>
+				</NavigationList>
+				<IconWrapper onClick={toggle}>
+					<X />
+					<VisuallyHidden>Close mobile menu</VisuallyHidden>
+				</IconWrapper>
+			</Panel>
+		</Menu>
 	);
 }
 
 const Menu = styled(Dialog)``;
 
-const Backdrop = styled(motion.div)`
+const Backdrop = styled.div`
 	position: absolute;
 	inset: 0;
-	background: hsl(0deg 0% 0%);
+	background: hsl(0deg 0% 0% / 0.4);
 `;
 
-const PanelWrapper = styled(motion.div)`
+const slideIn = keyframes`
+  0% {
+    transform: translateY(-100%);
+  }
+  100% {
+    transform: translateY(0);
+  }
+`;
+
+const slideOut = keyframes`
+  0% {
+    transform: translateY(0%);
+  }
+  100% {
+    transform: translateY(-100%);
+  }
+`;
+
+const Panel = styled(Dialog.Panel)`
 	position: absolute;
 	top: 0;
 	left: 0;
@@ -66,10 +79,7 @@ const PanelWrapper = styled(motion.div)`
 	padding: 68px 24px 52px 24px;
 	background: var(--color-gray-900);
 	color: var(--color-gray-100);
-`;
-
-const Panel = styled(Dialog.Panel)`
-	border: 1px solid var(--color-gray-900);
+	animation: ${props => (props.isOpen ? slideIn : slideOut)} 200ms ease-out both;
 `;
 
 const Logo = styled.span`
@@ -110,15 +120,6 @@ const IconWrapper = styled(UnstyledButton)`
 	padding: var(--padding);
 	margin-left: calc(var(--padding) * -1);
 	margin-right: calc(var(--padding) * -1);
-`;
-
-const Test = styled(motion.div)`
-	position: absolute;
-	bottom: 0;
-	left: 0;
-	right: 0;
-	height: 200px;
-	background: red;
 `;
 
 export default MobileMenu;
